# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DIMO Devices API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/vehicle/{tokenID}/commands/charge/start": {
            "post": {
                "description": "Start the vehicle charging.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device",
                    "integration",
                    "command"
                ],
                "summary": "Start the vehicle charging.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "tokenID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.CommandResponse"
                        }
                    }
                }
            }
        },
        "/vehicle/{tokenID}/commands/charge/stop": {
            "post": {
                "description": "Stop the vehicle charging.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device",
                    "integration",
                    "command"
                ],
                "summary": "Stop the vehicle charging.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "tokenID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.CommandResponse"
                        }
                    }
                }
            }
        },
        "/vehicle/{tokenID}/commands/doors/lock": {
            "post": {
                "description": "Lock the device's doors.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device",
                    "integration",
                    "command"
                ],
                "summary": "Lock the device's doors",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "tokenID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.CommandResponse"
                        }
                    }
                }
            }
        },
        "/vehicle/{tokenID}/commands/doors/unlock": {
            "post": {
                "description": "Unlock the device's doors.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device",
                    "integration",
                    "command"
                ],
                "summary": "Unlock the device's doors",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "tokenID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.CommandResponse"
                        }
                    }
                }
            }
        },
        "/vehicle/{tokenID}/commands/frunk/open": {
            "post": {
                "description": "Open the device's front trunk. Currently, this only works for Teslas connected through Tesla.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device",
                    "integration",
                    "command"
                ],
                "summary": "Open the device's front trunk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "tokenID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.CommandResponse"
                        }
                    }
                }
            }
        },
        "/vehicle/{tokenID}/commands/trunk/open": {
            "post": {
                "description": "Open the device's front trunk. Currently, this only works for Teslas connected through Tesla.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device",
                    "integration",
                    "command"
                ],
                "summary": "Open the device's rear trunk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "tokenID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.CommandResponse"
                        }
                    }
                }
            }
        },
...
        "/vehicle/{tokenId}/vin": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "updates the VIN on the user device record. Can optionally also update the protocol and the country code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-devices"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "token id",
                        "name": "tokenId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "VIN",
                        "name": "vin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.UpdateVINReq"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_controllers.CommandResponse": {
            "type": "object",
            "properties": {
                "requestId": {
                    "type": "string"
                }
            }
        },
...
        "internal_controllers.UpdateVINReq": {
            "type": "object",
            "required": [
                "vin"
            ],
            "properties": {
                "canProtocol": {
                    "description": "CANProtocol optional. Numeric style made up protocol. 6 = CAN11_500, 7 = CAN29_500, 66/77 are some UDS thing etc",
                    "type": "string"
                },
                "countryCode": {
                    "description": "CountryCode optional. Is set on the user device record",
                    "type": "string"
                },
                "signature": {
                    "description": "Signature is the hex-encoded result of the AutoPi signing the VIN. It must\nbe present to verify the VIN.",
                    "type": "string",
                    "example": "16b15f88bbd2e0a22d1d0084b8b7080f2003ea83eab1a00f80d8c18446c9c1b6224f17aa09eaf167717ca4f355bb6dc94356e037edf3adf6735a86fc3741f5231b"
                },
                "vin": {
                    "description": "VIN is a vehicle identification number. At the very least, it must be\n17 characters in length and contain only letters and numbers.",
                    "type": "string",
                    "example": "4Y1SL65848Z411439"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}